initial 작업 batch함수로 분리(main.js 코드 간소화)
