+++ scripts/util.js
@@ -1,4 +1,4 @@
-export function getCurrentDateString() {
+function getCurrentDateString() {
     let kr_days = ["일","월","화","수","목","금","토"];
     let currentDate = new Date();
     let year = currentDate.getFullYear();
@@ -6,4 +6,9 @@ export function getCurrentDateString() {
     let date = currentDate.getDate();
     let day = currentDate.getDay();
     return `${year}. ${month}. ${date}. ${kr_days[day]}요일`;
+}
+
+export function setUpCurrentTime() {
+    const currentDateString = getCurrentDateString();
+    document.getElementById("date_section").innerHTML = currentDateString;
 }
